Move read_bytes_from_finite_reader to decode
Move the `read_bytes_from_finite_reader` function to the `decode` module
and make it `pub(crate)`. Also move the associated
`ReadBytesFromFiniteReaderOpts`, in `decode` put it below where its used.
