+++ bitcoin/src/consensus/encode.rs
@@ -24,7 +24,8 @@ use crate::bip152::{PrefilledTransaction, ShortId};
 use crate::blockdata::block;
 use crate::blockdata::transaction::{Transaction, TxIn, TxOut};
 // TODO(Tobin): Remove this re-export once consensus refactor is done.
-pub use crate::consensus::decode::{Decodable, Error, ReadExt, MAX_VEC_SIZE};
+pub use crate::consensus::decode::{self, Decodable, Error, ReadExt, MAX_VEC_SIZE};
+use crate::consensus::decode::{read_bytes_from_finite_reader, ReadBytesFromFiniteReaderOpts};
 use crate::hash_types::{BlockHash, FilterHash, FilterHeader, TxMerkleNode};
 use crate::io::{self, Cursor};
 #[cfg(feature = "std")]
@@ -464,36 +465,6 @@ pub(crate) fn consensus_encode_with_size<W: io::Write + ?Sized>(
     Ok(vi_len + data.len())
 }
 
-struct ReadBytesFromFiniteReaderOpts {
-    len: usize,
-    chunk_size: usize,
-}
-
-/// Read `opts.len` bytes from reader, where `opts.len` could potentially be malicious.
-///
-/// This function relies on reader being bound in amount of data
-/// it returns for OOM protection. See [`Decodable::consensus_decode_from_finite_reader`].
-#[inline]
-fn read_bytes_from_finite_reader<D: io::Read + ?Sized>(
-    d: &mut D,
-    mut opts: ReadBytesFromFiniteReaderOpts,
-) -> Result<Vec<u8>, Error> {
-    let mut ret = vec![];
-
-    assert_ne!(opts.chunk_size, 0);
-
-    while opts.len > 0 {
-        let chunk_start = ret.len();
-        let chunk_size = core::cmp::min(opts.len, opts.chunk_size);
-        let chunk_end = chunk_start + chunk_size;
-        ret.resize(chunk_end, 0u8);
-        d.read_slice(&mut ret[chunk_start..chunk_end])?;
-        opts.len -= chunk_size;
-    }
-
-    Ok(ret)
-}
-
 impl Encodable for Vec<u8> {
     #[inline]
     fn consensus_encode<W: io::Write + ?Sized>(&self, w: &mut W) -> Result<usize, io::Error> {
@@ -507,7 +478,7 @@ impl Decodable for Vec<u8> {
         let len = VarInt::consensus_decode(r)?.0 as usize;
         // most real-world vec of bytes data, wouldn't be larger than 128KiB
         let opts = ReadBytesFromFiniteReaderOpts { len, chunk_size: 128 * 1024 };
-        read_bytes_from_finite_reader(r, opts)
+        decode::read_bytes_from_finite_reader(r, opts)
     }
 }
